Update wordlist
Intended criteria for word stimuli was SUBTLEX(US) word frequency from 1 to 5. Old list had log frequency from 1 to 5, which included a lot of very low frequency words inappropriate for the task. The new list is considerably smaller in the number of words (~2000 to ~500), might be a problem for long multi-session studies without repeating the words.