Add start of Markdown notes on deploying to Google Cloud Platform
Adding a Markdown file for notes on deploying the experiment (or
any experiment) to Google Cloud Platform, including how to set up
the account, the reason for doing things on GCP in the first place,
the role of Python, etc. This is meant to be both a broader
step-by-step guide as well as a little tutorial introduction to the
whole thing so that, if you need to deploy anything to the App Engine
in the future, you can refer to this document (and, hopefully, build
upon it).

This also includes a shell oneliner to generate an HTML file using
`pandoc` from the Markdown file. The first argument is the input
file and the second argument is the output file.