Add initial apa6-style manuscript draft
Add an initial blank LaTeX document using the `apa6` document class,
as well as a short shell script for generating the manuscript and a
gitignore file so that the LaTeX build artifacts (including the final
PDF document!) do not get included in the repository. A copy of the
final PDF should be placed in a location for safekeeping when
being submitted.