Add Bib(La)Tex bibliography and scripts
Adding a bibtex bibliography file (sourcemem.bib), updating the
manuscript to read from that file (using the addbibresource command),
and updating the makepdf scripts to include bibliography compilation.
The makepdf script compiles (and fails) silently; the makepdf-err
script compiles with full errors.

Compilation of the bibliography requires the use of the `biber`
application, which should be included in the TeXLive distribution
(perhaps in the -bibtex-extras package).