Add practice block
Added a practice block to both the simultaneous and sequential html documents. The number of trials in the practice block is equal to those of the actual experimental blocks. Labels for data from these trials are labelled with a prefix "practice_[phase name]" to distinguish it from trials in the experiment proper.