Change recognition confidence mapping
Changed valid choices and instructions for recognition task. "OLD" keys were 4, 5,and 6 along a 6 point confidence scale, but this has been moved to 8, 9, and 0 so participants do not have to move their hands from the previous task (1 and 0 in a true/false response) to be bunched up on the left side of the keyboard.