Query session ID on entry and return either fresh session or old
Entry into the experiment now queries whether an entity already exists
with the given external ID in the datastore. If the entity with that
external ID does exist in the datastore, the previous session ID is
returned and the cookie is set, logging the observer back in. If not,
a new session ID is entered in to the datastore and a fresh client
session is made.

This resolves issue #6.