Normalise hitting angle
Response error, as computed by the presentation and recall plugins, was normalised to be between 0 and 2*pi. However, the stored hitting angle was is expressed between -pi and pi as the function "normalise_angle" was not used. This is not a big deal as normalising can be done afterwards in analysis, but it is neater to have angles be uniformly expressed in the raw dataset.

Also added a line in the data handling python script so that angular error is directly saved in the dataset, rather than calculating it later from hitting and target angles.