Redirect to external URL
Added an if statement at the end of session-complete, which is where participants go after they finish any experimentla session, that checks if the participant is REP or not. If REP, then the original code should do exactly the same as before, but if the user is not (presumably this means they are from Prolific) then they are redirected to an external URL. At the moment these are just wikipedia links, but the Prolific urls will be substituted in.