Fix session ID not updating
Figured out that the session ID is handed to main.py through the url defined at line 230 of the experiment template html. I thought the only difference between the s1, s2, and s3 versions of the former templates was the wordlists they sampled. Actually, it was also the session ID specified on this line. When i made all 10 sessions call the same experiment, it was simply posting s1 data over and over, overwriting the last entry, which is not what we want. There is probably a smarter way of doing this, by say having a check for the last session ID and counting up by one, but this way I can be sure I am not doing anything wrong.