Initial plotting functions
Plot response error models, and recentered fits. These functions use ggplot, which im unable to control the smoothing on the edges of the model density plots, resulting in a smoothing to zero beyond the range (-pi, pi), which leads to this bowing that doesn't look good. Need to fix this before final draft.