Add RT Quantile Confidence
"Have you tried plotting error estimates on the .9 quantile? You can generate these for an individual subject using a bootstrap method where you sample the RTs with replacement and generate around 1000 simulated datasets. You can then calculate the quantiles for each simulated dataset and either calculate the standard deviation (which is the standard error) or the middle 95% (which is the 95% confidence interval)" - Adam