Script to refit diffusion models with free eta
Learning from lessons of EXP_INT, I'm using a shared function for all the models (diffusion_intrusion_model), that takes in the bounds as an argument rather than having the bounds set inside the model function. The reason for this is so that parameters can be set at 0 without the model complaining about boundaries since these are set separately for each model one level above the actual fitting code.