Plots to address reviewers comments
Two new plots for the response to reviewers. The first simulates the spatiotemporal model, allowing for a mixture of high and low criterion target responding, which helps get the 0.7 and 0,9 RT quantiles for high accuracy responses, which were previously missed. The second is to demonstrate the temporal and spatial gradients of the fitted models (just plotting the exponential function with fitted parameter estimates).